restructure components
